pyproject.toml
[project]
name = "ontologiacomputazionale"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anthropic==0.84.0",
    "anyio==4.12.1",
    "asttokens==3.0.1",
    "attrs==25.4.0",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.5",
    "click==8.3.1",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "docstring-parser==0.17.0",
    "executing==2.2.1",
    "fonttools==4.62.0",
    "frozenlist==1.8.0",
    "gensim==4.4.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "instructor==1.14.5",
    "ipython==9.11.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.11.1",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpickle==4.1.1",
    "jsonpointer==3.0.0",
    "kiwisolver==1.5.0",
    "langchain-core==1.2.17",
    "langchain-text-splitters==1.1.1",
    "langsmith==0.7.14",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "multidict==6.7.1",
    "networkx==3.6.1",
    "node2vec==0.4.3",
    "numpy==2.4.3",
    "openai==2.26.0",
    "optuna>=4.7.0",
    "orjson==3.11.7",
    "packaging==26.0",
    "pandas==3.0.1",
    "parso==0.8.6",
    "pexpect==4.9.0",
    "pillow==12.1.1",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "pyvis==0.3.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "sentence-transformers>=5.3.0",
    "setuptools==82.0.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smart-open==7.5.1",
    "sniffio==1.3.1",
    "stack-data==0.6.3",
    "tenacity==9.1.4",
    "threadpoolctl==3.6.0",
    "torch>=2.10.0",
    "torch-geometric>=2.7.0",
    "tqdm==4.67.3",
    "traitlets==5.14.3",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "wcwidth==0.6.0",
    "wrapt==2.1.2",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zstandard==0.25.0",
]
